requirements.txt
# LLM API clients
openai>=1.0.0
anthropic>=0.18.0
google-genai>=1.0.0

# Environment variable loading
python-dotenv>=1.0.0
